fix: equip ruby bolts at end of fight
Equips ruby bolts if present at the end of the fight.  This makes banker smoother as currently if the the player has diamonds equipped the bank script searches for them when withdrawing items, causing the script to rebank.  This is creating a bad pattern I believe and I've tested this locally to work for me.